requirements.txt
langchain==0.3.18
langchain-openai==0.3.5
langchain-anthropic==0.3.7
langgraph==0.2.72
langmem==0.0.8
python-dotenv==1.0.1
langchain_ollama==0.3.0